Extract SwapQuoteService
