+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/SwapViewModel.kt
@@ -7,44 +7,31 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.core.IAdapterManager
-import io.horizontalsystems.bankwallet.core.IBalanceAdapter
 import io.horizontalsystems.bankwallet.modules.swap.SwapMainModule
 import io.horizontalsystems.marketkit.models.Token
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import java.math.BigDecimal
 
 class SwapViewModel(
-    private val swapProvidersManager: SwapProvidersManager,
-    private val adapterManager: IAdapterManager
+    private val swapQuoteService: SwapQuoteService
 ) : ViewModel() {
-    private val quoteLifetime = 30000L
-    private var amountIn: BigDecimal? = null
-    private var tokenIn: Token? = null
-    private var tokenOut: Token? = null
-    private var quoting = false
-    private var quotes: List<SwapProviderQuote> = listOf()
-    private var preferredProvider: SwapMainModule.ISwapProvider? = null
-    private var error: Throwable? = null
-    private var quote: SwapProviderQuote? = null
-    private var availableBalance: BigDecimal? = null
+
+    private var swapQuoteState = swapQuoteService.stateFlow.value
 
     var uiState: SwapUiState by mutableStateOf(
         SwapUiState(
-            amountIn = amountIn,
-            tokenIn = tokenIn,
-            tokenOut = tokenOut,
-            quoting = quoting,
+            amountIn = swapQuoteState.amountIn,
+            tokenIn = swapQuoteState.tokenIn,
+            tokenOut = swapQuoteState.tokenOut,
+            quoting = swapQuoteState.quoting,
             swapEnabled = isSwapEnabled(),
-            quotes = quotes,
-            preferredProvider = preferredProvider,
-            quoteLifetime = quoteLifetime,
-            quote = quote,
-            error = error,
-            availableBalance = availableBalance
+            quotes = swapQuoteState.quotes,
+            preferredProvider = swapQuoteState.preferredProvider,
+            quoteLifetime = swapQuoteState.quoteLifetime,
+            quote = swapQuoteState.quote,
+            error = swapQuoteState.error,
+            availableBalance = swapQuoteState.availableBalance
         )
     )
         private set
@@ -52,125 +39,65 @@ class SwapViewModel(
     private var quotingJob: Job? = null
     private var scheduleReQuoteJob: Job? = null
 
-    fun onEnterAmount(v: BigDecimal?) {
-        amountIn = v
-        preferredProvider = null
+    init {
+        viewModelScope.launch {
+            swapQuoteService.stateFlow.collect {
+                handleUpdatedSwapQuoteState(it)
+            }
+        }
+    }
 
-        runQuotation()
+    private fun handleUpdatedSwapQuoteState(swapQuoteState: SwapQuoteService.State) {
+        this.swapQuoteState = swapQuoteState
+
+        emitState()
+    }
+
+    fun onEnterAmount(v: BigDecimal?) {
+        swapQuoteService.setAmount(v)
     }
 
     fun onSelectTokenIn(token: Token) {
-        tokenIn = token
-        refreshAvailableBalance()
-        preferredProvider = null
-
-        runQuotation()
+        swapQuoteService.setTokenIn(token)
     }
 
     fun onSelectTokenOut(token: Token) {
-        tokenOut = token
-        preferredProvider = null
-
-        runQuotation()
+        swapQuoteService.setTokenOut(token)
     }
 
     fun onSwitchPairs() {
-        val tmpTokenIn = tokenIn
-
-        tokenIn = tokenOut
-        refreshAvailableBalance()
-        tokenOut = tmpTokenIn
-
-        amountIn = quote?.quote?.amountOut
-
-        runQuotation()
+        swapQuoteService.switchPairs()
     }
 
     fun onSelectQuote(quote: SwapProviderQuote) {
-        preferredProvider = quote.provider
-        this.quote = quote
-
-        emitState()
+        swapQuoteService.selectQuote(quote)
     }
 
     private fun emitState() {
         viewModelScope.launch {
             uiState = SwapUiState(
-                amountIn = amountIn,
-                tokenIn = tokenIn,
-                tokenOut = tokenOut,
-                quoting = quoting,
+                amountIn = swapQuoteState.amountIn,
+                tokenIn = swapQuoteState.tokenIn,
+                tokenOut = swapQuoteState.tokenOut,
+                quoting = swapQuoteState.quoting,
                 swapEnabled = isSwapEnabled(),
-                quotes = quotes,
-                preferredProvider = preferredProvider,
-                quoteLifetime = quoteLifetime,
-                quote = quote,
-                error = error,
-                availableBalance = availableBalance
+                quotes = swapQuoteState.quotes,
+                preferredProvider = swapQuoteState.preferredProvider,
+                quoteLifetime = swapQuoteState.quoteLifetime,
+                quote = swapQuoteState.quote,
+                error = swapQuoteState.error,
+                availableBalance = swapQuoteState.availableBalance
             )
         }
     }
 
-    private fun isSwapEnabled() = quote != null
-
-    private fun refreshAvailableBalance() {
-        availableBalance = tokenIn?.let {
-            (adapterManager.getAdapterForToken(it) as? IBalanceAdapter)?.balanceData?.available
-        }
-    }
-
-    private fun runQuotation() {
-        quotingJob?.cancel()
-        scheduleReQuoteJob?.cancel()
-
-        quotes = listOf()
-        quote = null
-        error = null
-
-        val amountIn = amountIn
-        val tokenIn = tokenIn
-        val tokenOut = tokenOut
-
-        if (amountIn == null || amountIn <= BigDecimal.ZERO || tokenIn == null || tokenOut == null) {
-            quoting = false
-            emitState()
-        } else {
-            quoting = true
-            emitState()
-
-            quotingJob = viewModelScope.launch(Dispatchers.Default) {
-                quotes = swapProvidersManager.getQuotes(tokenIn, tokenOut, amountIn).sortedByDescending { it.quote.amountOut }
-                if (preferredProvider != null && quotes.none { it.provider == preferredProvider}) {
-                    preferredProvider = null
-                }
-
-                if (quotes.isEmpty()) {
-                    error = SwapRouteNotFound()
-                } else {
-                    quote = preferredProvider
-                        ?.let { provider -> quotes.find { it.provider == provider } }
-                        ?: quotes.firstOrNull()
-                }
-
-                quoting = false
-                emitState()
-
-                scheduleReQuote()
-            }
-        }
-    }
-
-    private fun scheduleReQuote() {
-        scheduleReQuoteJob = viewModelScope.launch {
-            delay(quoteLifetime)
-            runQuotation()
-        }
-    }
+    private fun isSwapEnabled() = swapQuoteState.quote != null
 
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return SwapViewModel(SwapProvidersManager(), App.adapterManager) as T
+            val swapQuoteService = SwapQuoteService(SwapProvidersManager(), App.adapterManager)
+            return SwapViewModel(swapQuoteService) as T
         }
     }
 }
